feature fix :     - pathvariable 오류 수정     - QuizDto 날짜 형식 지정

+++ src/main/java/dreamdiary/quiz/domain/entity/QuizEntity.java
@@ -23,7 +23,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 public class QuizEntity {
   @GeneratedValue(strategy =  GenerationType.AUTO)
   @Id
-  private Long quizId;
+  private Integer quizId;
 
   @Column(nullable = false)
   private String title;
